refactor(1013): remove SitesByAPModelExporter facade (Cat B, position 28)
Extract SitesByAPModelExporter to src/export/sites_by_ap_model_exporter.py
using the SC-001 facade pattern. Class body in MistHelper.py replaced with
a comment breadcrumb; alphabetical re-export alias added to the imports
block so callers using MistHelper.SitesByAPModelExporter keep working.

- New: src/export/sites_by_ap_model_exporter.py (11 static methods,
  lazy mh = importlib.import_module("MistHelper") in helpers touching
  APICoreFetchUtils, InputUtils, DataExporter, ConfigUtils).
- MistHelper.py: -148 LOC class body -> 2-line breadcrumb comment.
- pyproject.toml: add coverage-omit entry for the new module.

Local gate: black/ruff clean, py_compile OK, MistHelper.py --test
66/0/131 baseline preserved.
